requirements.txt
asgiref==3.6.0
Django==4.2.1
pycryptodome==3.17
sqlparse==0.4.4
tzdata==2023.3
